refactor(quality): move state_contract.py into state/ package (P4)
Moves the orphan top-level module into the existing state/ package as
state/contract.py. state_contract.py becomes a thin backward-compat shim.

All 3 src/ callers and 3 test files updated to import from the canonical
location. state/__init__.py now re-exports the contract symbols.
